requirements.txt
# ── Step 1: Ingesta y Persistencia ────────────────────────────────────────
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.1
sqlalchemy==2.0.30
requests==2.32.3
tenacity==8.4.1
python-dateutil==2.9.0
python-dotenv==1.0.1
rich==13.7.1

# ── Step 2: Descarga de pliegos (Playwright) ──────────────────────────────
# playwright==1.44.0
# (ejecutar también: playwright install chromium)

# ── Step 3: Extracción de PDFs ────────────────────────────────────────────
# pymupdf==1.24.5
# google-generativeai==0.7.2

# ── Step 4: Análisis local ────────────────────────────────────────────────
# ollama==0.2.1

# ── Step 5: Notificaciones ────────────────────────────────────────────────
# openai==1.35.7
